requirements.txt
discord.py==2.4.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
